+++ src/main/java/com/songoda/epiclevels/players/EPlayer.java
@@ -23,7 +23,7 @@ public class EPlayer {
 
     private final UUID uuid;
 
-    private long experience;
+    private double experience;
 
     private int mobKills;
     private int playerKills;
@@ -35,7 +35,7 @@ public class EPlayer {
 
     private static ScriptEngine engine = null;
 
-    public EPlayer(UUID uuid, long experience, int mobKills, int playerKills, int deaths, int killstreak, int bestKillstreak) {
+    public EPlayer(UUID uuid, double experience, int mobKills, int playerKills, int deaths, int killstreak, int bestKillstreak) {
         this.uuid = uuid;
         this.experience = experience;
         this.mobKills = mobKills;
@@ -51,7 +51,7 @@ public class EPlayer {
     }
 
     public EPlayer(UUID uuid) {
-        this(uuid, Setting.START_EXP.getLong(), 0, 0, 0, 0, 0);
+        this(uuid, Setting.START_EXP.getDouble(), 0, 0, 0, 0, 0);
     }
 
     public UUID getUniqueId() {
@@ -62,7 +62,7 @@ public class EPlayer {
         return Bukkit.getOfflinePlayer(uuid);
     }
 
-    public long addExperience(long experience) {
+    public double addExperience(double experience) {
         if (experience == 0) return experience;
         EpicLevels plugin = EpicLevels.getInstance();
         if (experience < 0L) {
@@ -98,8 +98,8 @@ public class EPlayer {
                 for (Player pl : Bukkit.getOnlinePlayers().stream().filter(p -> p != player).collect(Collectors.toList()))
                     pl.sendMessage(plugin.getReferences().getPrefix() + plugin.getLocale().getMessage("event.levelup.announcement", player.getName(), getLevel()));
         }
-        if (this.experience > Setting.MAX_EXP.getLong())
-            this.experience = Setting.MAX_EXP.getLong();
+        if (this.experience > Setting.MAX_EXP.getDouble())
+            this.experience = Setting.MAX_EXP.getDouble();
         return this.experience;
     }
 
@@ -120,7 +120,7 @@ public class EPlayer {
         return multiplier;
     }
 
-    public long getExperience() {
+    public double getExperience() {
         return experience;
     }
